+++ vasuki/core.py
@@ -38,8 +38,7 @@ class VasukiCommand:
             result = generate_ulid()
 
         elif self.options.naga19:
-            assert self.options.size is None, 'Option "size" makes no sense for Nagamani19'
-            result = generate_nagamani19()
+            result = generate_nagamani19(size=self.options.size)
 
         elif self.options.gibberish:
             result = generate_gibberish(self.options.size)