+++ vasuki/unique/nagamani.py
@@ -1,26 +1,20 @@
 # -*- coding: utf-8 -*-
 # (c) 2019 Andreas Motl <andreas@terkin.org>
 # Apache License, Version 2
-import os
 from datetime import datetime
-from binascii import hexlify
 from hashids import Hashids
 """
-Nagamani19, a short, unique, non-sequential identifier based on Hashids.
+Nagamani - short, unique, non-sequential identifiers based on Hashids.
 """
 
-salt = None
+precisions = {
+    's': 1,
+    'ms': 1000,
+    'ns': 1000000,
+}
 
 
-def generate_nagamani19():
-    # https://community.hiveeyes.org/t/gestaltung-der-anonymen-identifizierer-ids-fur-die-adressierung-von-imkern-messknoten-und-anderen-entitaten/1079/8
-    global salt
-    if salt is None:
-        salt = gensalt()
-    return naga_id(2019, salt)
-
-
-def naga_id(year, salt):
+def nagamani_id(year, salt, precision):
     """
     Generate unique ids based on Hashids.
 
@@ -39,8 +33,16 @@ def naga_id(year, salt):
         yzgOlvap
 
     """
-    now = int(round((datetime.utcnow() - datetime(year, 1, 1)).total_seconds() * 1000))
-    return hashify(salt, now)
+
+    assert type(year) is int, 'Year must be integer'
+    assert salt is not None, 'Salt must be given'
+    assert precision is not None, 'Precision must be one of s, ms, ns'
+
+    scaling = precisions.get(precision)
+
+    duration = datetime.utcnow() - datetime(year, 1, 1)
+    duration_scaled = int(duration.total_seconds() * scaling)
+    return hashify(salt, duration_scaled)
 
 
 def hashify(salt, *data):
@@ -59,20 +61,3 @@ def hashify(salt, *data):
     """
     hashids = Hashids(salt=salt)
     return hashids.encode(*data)
-
-
-def gensalt():
-    """
-    This generates a salt from 24 random bytes from an OS-specific randomness source.
-    The returned data should be unpredictable enough for cryptographic applications,
-    though its exact quality depends on the OS implementation.
-
-    https://docs.python.org/3/library/os.html#os.urandom
-
-    Examples::
-
-        b5f95cead701f2488d5668decb0d63a30e7ddb4c21f26574
-        b4157e5459c88a6c454186492ee629ca097f8a60cbfb1a36
-        de1ba437524e540e3b0d55617afcad5677b982d9e1f45f9d
-    """
-    return hexlify(os.urandom(24)).decode()