+++ CHANGES.rst
@@ -20,6 +20,7 @@ in progress
 - Wording: Rename ``--wordlength`` to ``--size``
 - Swap display order of generators
 - Generate multiple tokens at once
+- Make Nagamani19 generator obtain size, mapping it to precision
 
 
 2019-06-28 0.0.0